requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.2.0
beautifulsoup4==4.14.2
bleach==6.2.0
brotli==1.2.0
certifi==2025.8.3
cffi==1.16.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastapi==0.121.0
fastjsonschema==2.21.2
ffmpy==0.6.4
filelock==3.20.0
fonttools==4.60.0
fqdn==1.5.1
fsspec==2025.10.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mpmath==1.3.0
mypysql==0.1.7
mysql-connector-python==9.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.0.1
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
polars==1.35.1
polars-runtime-32==1.35.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2024.1
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.27.1
ruff==0.14.4
safehttpx==0.1.7
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.31
sshtunnel==0.4.0
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
terminado==0.18.1
tinycss2==1.4.0
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.226
ultralytics-thop==2.0.18
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
